+++ src/bin_dal_node/configuration_file.ml
@@ -80,6 +80,7 @@ type t = {
   service_name : string option;
   service_namespace : string option;
   experimental_features : experimental_features;
+  fetch_trusted_setup : bool;
 }
 
 let default_data_dir = Filename.concat (Sys.getenv "HOME") ".tezos-dal-node"
@@ -121,6 +122,8 @@ let default_experimental_features =
   let default_sqlite3_backend = false in
   {sqlite3_backend = default_sqlite3_backend}
 
+let default_fetch_trusted_setup = true
+
 let default =
   {
     data_dir = default_data_dir;
@@ -139,6 +142,7 @@ let default =
     service_name = None;
     service_namespace = None;
     experimental_features = default_experimental_features;
+    fetch_trusted_setup = default_fetch_trusted_setup;
   }
 
 let neighbor_encoding : neighbor Data_encoding.t =
@@ -193,6 +197,7 @@ let encoding : t Data_encoding.t =
            service_name;
            service_namespace;
            experimental_features;
+           fetch_trusted_setup;
          } ->
       ( ( data_dir,
           rpc_addr,
@@ -209,7 +214,8 @@ let encoding : t Data_encoding.t =
           version,
           service_name,
           service_namespace,
-          experimental_features ) ))
+          experimental_features,
+          fetch_trusted_setup ) ))
     (fun ( ( data_dir,
              rpc_addr,
              listen_addr,
@@ -225,7 +231,8 @@ let encoding : t Data_encoding.t =
              version,
              service_name,
              service_namespace,
-             experimental_features ) ) ->
+             experimental_features,
+             fetch_trusted_setup ) ) ->
       {
         data_dir;
         rpc_addr;
@@ -243,6 +250,7 @@ let encoding : t Data_encoding.t =
         service_name;
         service_namespace;
         experimental_features;
+        fetch_trusted_setup;
       })
     (merge_objs
        (obj10
@@ -296,7 +304,7 @@ let encoding : t Data_encoding.t =
              ~description:"The point for the DAL node metrics server"
              (Encoding.option P2p_point.Id.encoding)
              None))
-       (obj6
+       (obj7
           (dft
              "history_mode"
              ~description:"The history mode for the DAL node"
@@ -322,7 +330,12 @@ let encoding : t Data_encoding.t =
              "experimental_features"
              ~description:"Experimental features"
              experimental_features_encoding
-             default_experimental_features)))
+             default_experimental_features)
+          (dft
+             "fetch_trusted_setup"
+             ~description:"Install trusted setup"
+             bool
+             true)))
 
 module V0 = struct
   type t = {
@@ -476,6 +489,7 @@ let from_v0 v0 =
     service_name = None;
     service_namespace = None;
     experimental_features = default_experimental_features;
+    fetch_trusted_setup = true;
   }
 
 type error += DAL_node_unable_to_write_configuration_file of string