+++ src/bin_dal_node/configuration_file.mli
@@ -66,6 +66,8 @@ type t = {
       (** Name of the service provided by this node. *)
   service_namespace : string option;  (** Namespace for the service. *)
   experimental_features : experimental_features;  (** Experimental features.  *)
+  fetch_trusted_setup : bool;
+      (** Should the trusted setup be downloaded if not found or has invalid hash. *)
 }
 
 (** [default] is the default configuration. *)